refactor(creators): create separate create note file
