+++ package/src/creators/create-root.ts
@@ -1,8 +1,6 @@
-import { Route } from '@angular/router';
 import { enhance } from '../utils/enhance';
 import { hub, nextHubValue } from '../hub';
-import { setRouteName } from '../utils/name';
-import { Hub, Note, Notes, RouteNameOptions, Slice } from '../interfaces';
+import { Hub, Notes, Slice } from '../interfaces';
 
 /**
  * Creates main parent routes
@@ -26,30 +24,3 @@ export function createRoot<R, C = {}>(
 
   return hub.value[name];
 }
-
-/**
- * creates a route note
- */
-export const createNote = <R = {}, C = {}>(
-  routes: Route[],
-  nameOptions: RouteNameOptions = {}
-): R & C =>
-  routes.reduce(
-    (acc: R & C, route: Route): R & C => {
-      const note: Note = {
-        path: route.path,
-        name: route['name'] || setRouteName(route.path, nameOptions)
-      };
-
-      if (route.loadChildren) {
-        note.lazy = route.loadChildren;
-      }
-
-      if (route.children) {
-        note.children = createNote(route.children, nameOptions);
-      }
-
-      return { ...acc, [note.name]: note };
-    },
-    {} as R & C
-  );
