+++ package/src/creators/index.ts
@@ -1,3 +1,4 @@
-export { createRoot, createNote } from './create-root';
+export { createRoot } from './create-root';
 export { createFeature } from './create-feature';
 export { createUnion } from './create-union';
+export { createNote } from './create-note';
